aifoundry/functions/requirements.txt
azure-functions
azure-ai-projects
azure-identity
azure-ai-contentsafety
requests
